projeto final pronto

+++ Estudio/FrmCadastrarModalidade.cs
@@ -36,7 +36,7 @@ namespace Estudio
                 {
                     MessageBox.Show("cadastro realizado com sucesso", "Alerta do Sistema", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
-                Modalidade mod = new Modalidade(txtDescricao.Text, float.Parse(txtAlunos.Text), int.Parse(txtAulas.Text), int.Parse(txtPreco.Text));
+                Modalidade mod = new Modalidade(txtDescricao.Text, int.Parse(txtAlunos.Text),int.Parse(txtAulas.Text), int.Parse(txtPreco.Text));
                 if (mod.cadastrarModalidade())
                 {
                     MessageBox.Show("Modalidade cadastrada");
